# refactor(stdlib/universe): convert last to index selector