I was looking at the implementation of first and last, and I noticed
that first was an index selector while last was a row selector. It
seemed strange that they weren't a nearly identical implementation. Now
they are both index selectors, which improved the benchmarks for last:

```
name    old time/op    new time/op    delta
Last-4     395ns ± 6%     262ns ±13%  -33.61%  (p=0.000 n=8+8)

name    old alloc/op   new alloc/op   delta
Last-4      224B ± 0%      185B ± 0%  -17.41%  (p=0.000 n=8+8)

name    old allocs/op  new allocs/op  delta
Last-4      9.00 ± 0%      2.00 ± 0%  -77.78%  (p=0.000 n=8+8)
```