@nathanielc 2019-05-08
@jlapacik @mark-rushakoff I think this breaks the implementation of last. With the change it selects the last row per flux.ColReader instead of the last row of the flux.Table. It looks like the test cases only have a single flux.ColReader per table so it wasn't caught. 

Can either of you look into adding a test case that uses multiple flux.ColReaders to ensure that this works? I could be remembering the internals wrong, but we need the test case anyways.