@jlapacik 2019-05-08
I created the following issue #1260 and am looking into it.