@jlapacik 2019-05-08
#1264 updates the tests for `last` as well as other selectors (`max` and `min`) that only used a single column reader per table. It also reverts this commit which unfortunately breaks the implementation of `last` when multiple column readers are needed/used to read a flux table.